requirements.txt
Flask==3.0.3
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
werkzeug>=3.0.0
gunicorn>=21.2.0
supabase>=2.0.0
python-dotenv==1.0.0
optuna>=3.4.0
safetensors>=0.4.0
